refactor: 영화 데이터소스를 공통 모델로 빼고 첫 화면에서 영화를 선택하면 그 리스트뷰의 id를 intent로 넘겨주는 방식으로 변경
