+++ app/src/main/java/woowacourse/movie/reservation/model/MovieReservationData.kt
@@ -1,0 +1,5 @@
+package woowacourse.movie.reservation.model
+
+object MovieReservationData {
+    var movieId: Long = -1
+}
