--- app/src/main/java/woowacourse/movie/reservation/model/MovieReservationMovieData.kt
@@ -1,8 +1,0 @@
-package woowacourse.movie.reservation.model
-
-import woowacourse.movie.list.model.Movie
-import woowacourse.movie.list.model.MovieDataSource.emptyMovie
-
-object MovieReservationMovieData {
-    var movieData: Movie = emptyMovie
-}
