+++ app/src/main/java/woowacourse/movie/ticket/contract/MovieTicketContract.kt
@@ -2,11 +2,10 @@ package woowacourse.movie.ticket.contract
 
 import woowacourse.movie.reservation.model.Count
 import woowacourse.movie.ticket.model.Ticket
-import woowacourse.movie.ticket.presenter.MovieTicketPresenter
 
 interface MovieTicketContract {
     interface View {
-        val presenter: MovieTicketPresenter
+        val presenter: Presenter
 
         fun showTicketView(
             info: Ticket,
